pyproject.toml
[project]
name = "InvestAPI"
version = "0.1.0"
description = "API for gathering Stock, ETFs, Crypto and Steam item prices"
readme = "README.md"
authors = [
    { name = "Bibikau Maxim", email = "bib.maxim@gmail.com" }
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "yfinance>=0.2.65"
]

[project.scripts]
investapi = "investapi.app:main"

[project.urls]
Homepage = "https://github.com/Max2772/InvestAPI"

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = [
    "investapi",
]
